refactor(hooks): shared useFilesystemProviders hook
Two consumers (useHasFilesystemProviders, FilesystemPanel) used the
same queryKey with different staleTime. Per-observer staleTime means
the Panel's default-0 defeated the Hook's 5_000 dedup whenever the
Panel was open.

Extract one hook owning both query + socket invalidation.
